Added player implementation.
* Entity.cs, Button.cs, State.cs, MenuState.cs, DifficultyState.cs:
  Modified to use float instead of double for delta time.

* Player.cs:
  Initial implementation.

* Game.cs:
  Cast DeltaTime to float from double.

* Input.cs:
  Cleanup.

* GameState.cs:
  Added EndGame().
  Added creation / reset code for player.
  Modified to use float instead of double for delta time.

* Extensions.cs:
  Added MathUtil.Clamp.
  Added MathUtil.Lerp.

+++ SpaceTapper/Source/Entity.cs
@@ -16,10 +16,10 @@ namespace SpaceTapper
 			State    = state;
 		}
 
-		protected abstract void UpdateSelf(double dt);
+		protected abstract void UpdateSelf(float dt);
 		protected abstract void DrawSelf(RenderTarget target, RenderStates states);
 
-		public void Update(double dt)
+		public void Update(float dt)
 		{
 			UpdateSelf(dt);
